debian/rules: use gcc-4.8, g++-4.8 (XXX)
explicitly build with gcc/g++ 4.8. this is (currently) the default gcc
version on all but a few debian architectures. on some architectures the
default gcc is too old to compile the __atomic*() builtins, which were
introduced in gcc 4.7.

remember to remove this hack once the default gcc version on all debian
architectures is >= 4.7.